+++ src/Traits/InputTrait.php
@@ -46,6 +46,19 @@ trait InputTrait
         return $this->toHtmlString('<input'.$this->html->attributes($options).'>');
     }
 
+    /** Create a search input field.
+     *
+     * @param  string $name
+     * @param  string $value
+     * @param  array  $options
+     *
+     * @return \Illuminate\Support\HtmlString
+     */
+    public function search($name, $value = null, $options = [])
+    {
+        return $this->input('search', $name, $value, $options);
+    }
+
     /**
      * Create a text input field.
      *
@@ -268,7 +281,7 @@ trait InputTrait
         // the element. Then we'll create the final textarea elements HTML for us.
         $options = $this->html->attributes($options);
 
-        return $this->toHtmlString('<textarea'.$options.'>'.$this->entities($value, true).'</textarea>');
+        return $this->toHtmlString('<textarea'.$options.'>'.$this->entities($value).'</textarea>');
     }
 
     /**