+++ tests/FormAccessibleTest.php
@@ -1,18 +1,19 @@
 <?php
 
+use Mockery as m;
 use Carbon\Carbon;
-use Collective\Html\Eloquent\FormAccessible;
-use Illuminate\Database\Eloquent\Model;
+use Illuminate\Http\Request;
+use PHPUnit\Framework\TestCase;
 use Collective\Html\FormBuilder;
 use Collective\Html\HtmlBuilder;
-use Illuminate\Contracts\View\Factory;
-use Illuminate\Http\Request;
-use Illuminate\Routing\RouteCollection;
 use Illuminate\Routing\UrlGenerator;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Routing\RouteCollection;
+use Collective\Html\Eloquent\FormAccessible;
 use Illuminate\Database\Capsule\Manager as Capsule;
-use Mockery as m;
 
-class FormAccessibleTest extends PHPUnit_Framework_TestCase
+class FormAccessibleTest extends TestCase
 {
     public function setUp()
     {
@@ -27,6 +28,8 @@ class FormAccessibleTest extends PHPUnit_Framework_TestCase
           'address'    => [
               'street' => 'abcde st'
           ],
+          'array'      => [1, 2, 3,],
+          'transform_key' => 'testing testing',
           'created_at' => $this->now,
           'updated_at' => $this->now,
         ];
@@ -46,10 +49,50 @@ class FormAccessibleTest extends PHPUnit_Framework_TestCase
         $this->assertEquals($model->getFormValue('created_at'), $this->now->timestamp);
     }
 
+    public function testItCanMutateRelatedValuesForForms()
+    {
+        $model = new ModelThatUsesForms($this->modelData);
+        $relatedModel = new ModelThatUsesForms($this->modelData);
+        $relatedModel->address = [
+            'street' => '123 Evergreen Terrace'
+        ];
+        $model->setRelation('related', $relatedModel);
+
+        $this->formBuilder->setModel($model);
+
+        $this->assertEquals($this->formBuilder->getValueAttribute('related[string]'), 'ponmlkjihgfedcba');
+        $this->assertEquals($this->formBuilder->getValueAttribute('related[address][street]'), '123 Evergreen Terrace');
+    }
+
     public function testItCanGetRelatedValueForForms()
     {
         $model = new ModelThatUsesForms($this->modelData);
         $this->assertEquals($model->getFormValue('address.street'), 'abcde st');
+    }
+    
+    public function testItCanUseGetAccessorValuesWhenThereAreNoFormAccessors()
+    {
+        $model = new ModelThatUsesForms($this->modelData);
+        $this->formBuilder->setModel($model);
+        
+        $this->assertEquals($this->formBuilder->getValueAttribute('email'), 'mutated@tjshafer.com');
+    }
+
+    public function testItReturnsSameResultWithAndWithoutThisFeature()
+    {
+        $modelWithAccessor = new ModelThatUsesForms($this->modelData);
+        $modelWithoutAccessor = new ModelThatDoesntUseForms($this->modelData);
+
+        $this->formBuilder->setModel($modelWithAccessor);
+        $valuesWithAccessor[] = $this->formBuilder->getValueAttribute('array');
+        $valuesWithAccessor[] = $this->formBuilder->getValueAttribute('array[0]');
+        $valuesWithAccessor[] = $this->formBuilder->getValueAttribute('transform.key');
+        $this->formBuilder->setModel($modelWithoutAccessor);
+        $valuesWithoutAccessor[] = $this->formBuilder->getValueAttribute('array');
+        $valuesWithoutAccessor[] = $this->formBuilder->getValueAttribute('array[0]');
+        $valuesWithoutAccessor[] = $this->formBuilder->getValueAttribute('transform.key');
+
+        $this->assertEquals($valuesWithAccessor, $valuesWithoutAccessor);
     }
 
     public function testItCanStillMutateValuesForViews()
@@ -95,6 +138,11 @@ class ModelThatUsesForms extends Model
     public function getCreatedAtAttribute($value)
     {
         return '1 second ago';
+    }
+
+    public function getEmailAttribute($value)
+    {
+        return 'mutated@tjshafer.com';
     }
 }