+++ tests/HtmlBuilderTest.php
@@ -1,13 +1,14 @@
 <?php
 
-use Illuminate\Contracts\View\Factory;
-use Collective\Html\HtmlBuilder;
-use Illuminate\Http\Request;
-use Illuminate\Routing\RouteCollection;
-use Illuminate\Routing\UrlGenerator;
 use Mockery as m;
+use Illuminate\Http\Request;
+use PHPUnit\Framework\TestCase;
+use Collective\Html\HtmlBuilder;
+use Illuminate\Routing\UrlGenerator;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Routing\RouteCollection;
 
-class HtmlBuilderTest extends PHPUnit_Framework_TestCase
+class HtmlBuilderTest extends TestCase
 {
 
     /**
@@ -47,7 +48,7 @@ class HtmlBuilderTest extends PHPUnit_Framework_TestCase
 
         $ol = $this->htmlBuilder->ol($list, $attributes);
 
-        $this->assertEquals('<ol class="example"><li>foo</li><li>bar</li><li>&amp;amp;</li></ol>', $ol);
+        $this->assertEquals('<ol class="example"><li>foo</li><li>bar</li><li>&amp;</li></ol>', $ol);
     }
 
     public function testUl()
@@ -58,7 +59,7 @@ class HtmlBuilderTest extends PHPUnit_Framework_TestCase
 
         $ul = $this->htmlBuilder->ul($list, $attributes);
 
-        $this->assertEquals('<ul class="example"><li>foo</li><li>bar</li><li>&amp;amp;</li></ul>', $ul);
+        $this->assertEquals('<ul class="example"><li>foo</li><li>bar</li><li>&amp;</li></ul>', $ul);
     }
 
     public function testMeta()